Move shared functionality from master_controller.go and node_controller.go to external_controller.go. The only functional difference is updating status, which is implemented by passing updatePolicyStatusFunc
