+++ go-controller/pkg/ovn/controller/apbroute/external_controller.go
@@ -1,22 +1,37 @@
 package apbroute
 
 import (
+	"errors"
 	"fmt"
+	"reflect"
 	"strings"
 	"sync"
-
-	adminpolicybasedrouteapi "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/crd/adminpolicybasedroute/v1"
-	adminpolicybasedroutelisters "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/crd/adminpolicybasedroute/v1/apis/listers/adminpolicybasedroute/v1"
-	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/ovn/controller/apbroute/gateway_info"
-	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/syncmap"
+	"time"
 
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
 	ktypes "k8s.io/apimachinery/pkg/types"
+	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	"k8s.io/apimachinery/pkg/util/sets"
+	"k8s.io/apimachinery/pkg/util/wait"
+	coreinformers "k8s.io/client-go/informers/core/v1"
 	corev1listers "k8s.io/client-go/listers/core/v1"
+	"k8s.io/client-go/tools/cache"
+	"k8s.io/client-go/util/workqueue"
 	"k8s.io/klog/v2"
+
+	nettypes "github.com/k8snetworkplumbingwg/network-attachment-definition-client/pkg/apis/k8s.cni.cncf.io/v1"
+	adminpolicybasedrouteapi "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/crd/adminpolicybasedroute/v1"
+	adminpolicybasedrouteinformer "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/crd/adminpolicybasedroute/v1/apis/informers/externalversions/adminpolicybasedroute/v1"
+	adminpolicybasedroutelisters "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/crd/adminpolicybasedroute/v1/apis/listers/adminpolicybasedroute/v1"
+	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/factory"
+	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/ovn/controller/apbroute/gateway_info"
+	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/syncmap"
+)
+
+const (
+	maxRetries = 15
 )
 
 func insertSet(s1, s2 sets.Set[string]) {
@@ -178,24 +193,37 @@ type routePolicyConfig struct {
 
 type externalPolicyManager struct {
 	stopCh <-chan struct{}
-	// route policies
-	routeLister adminpolicybasedroutelisters.AdminPolicyBasedExternalRouteLister
-	// Pods
-	podLister corev1listers.PodLister
-	// Namespaces
-	namespaceLister corev1listers.NamespaceLister
+
 	// policyReferencedObjects should only be accessed with policyReferencedObjectsLock
 	policyReferencedObjectsLock sync.RWMutex
 	// policyReferencedObjects is a cache of objects every policy has selected for its config.
 	// With this cache namespace and pod handlers may fetch affected policies for cleanup.
 	// key is policyName.
 	policyReferencedObjects map[string]*policyReferencedObjects
+
 	// routePolicySyncCache is a cache of configures states for policies, key is policyName.
 	routePolicySyncCache *syncmap.SyncMap[*routePolicyState]
 	// networkClient is an interface that exposes add and delete GW IPs. There are 2 structs that implement this contract: one to interface with the north bound DB and another one for the conntrack.
 	// the north bound is used by the master controller to add and delete the logical static routes, whilst the conntrack is used by the node controller to ensure that the ECMP entries are removed
 	// when a gateway IP is no longer an egress access point.
 	netClient networkClient
+
+	// route policies
+	routeLister   adminpolicybasedroutelisters.AdminPolicyBasedExternalRouteLister
+	routeInformer cache.SharedIndexInformer
+	routeQueue    workqueue.RateLimitingInterface
+
+	// Pods
+	podLister   corev1listers.PodLister
+	podInformer cache.SharedIndexInformer
+	podQueue    workqueue.RateLimitingInterface
+
+	// Namespaces
+	namespaceQueue    workqueue.RateLimitingInterface
+	namespaceLister   corev1listers.NamespaceLister
+	namespaceInformer cache.SharedIndexInformer
+
+	updatePolicyStatusFunc func(policyName string, gwIPs sets.Set[string], processedError error) error
 }
 
 type policyReferencedObjects struct {
@@ -206,23 +234,395 @@ type policyReferencedObjects struct {
 
 func newExternalPolicyManager(
 	stopCh <-chan struct{},
-	podLister corev1listers.PodLister,
-	namespaceLister corev1listers.NamespaceLister,
-	routeLister adminpolicybasedroutelisters.AdminPolicyBasedExternalRouteLister,
-	netClient networkClient) *externalPolicyManager {
+	podInformer coreinformers.PodInformer,
+	namespaceInformer coreinformers.NamespaceInformer,
+	apbRouteInformer adminpolicybasedrouteinformer.AdminPolicyBasedExternalRouteInformer,
+	netClient networkClient,
+	updatePolicyStatusFunc func(policyName string, gwIPs sets.Set[string], processedError error) error) *externalPolicyManager {
 
 	m := externalPolicyManager{
 		stopCh:                      stopCh,
-		routeLister:                 routeLister,
-		podLister:                   podLister,
-		namespaceLister:             namespaceLister,
 		policyReferencedObjectsLock: sync.RWMutex{},
 		policyReferencedObjects:     map[string]*policyReferencedObjects{},
 		routePolicySyncCache:        syncmap.NewSyncMap[*routePolicyState](),
 		netClient:                   netClient,
+
+		routeLister:   apbRouteInformer.Lister(),
+		routeInformer: apbRouteInformer.Informer(),
+		routeQueue: workqueue.NewNamedRateLimitingQueue(
+			workqueue.NewItemFastSlowRateLimiter(time.Second, 5*time.Second, 5),
+			"adminpolicybasedexternalroutes",
+		),
+		podLister:   podInformer.Lister(),
+		podInformer: podInformer.Informer(),
+		podQueue: workqueue.NewNamedRateLimitingQueue(
+			workqueue.NewItemFastSlowRateLimiter(time.Second, 5*time.Second, 5),
+			"apbexternalroutepods",
+		),
+		namespaceLister:   namespaceInformer.Lister(),
+		namespaceInformer: namespaceInformer.Informer(),
+		namespaceQueue: workqueue.NewNamedRateLimitingQueue(
+			workqueue.NewItemFastSlowRateLimiter(time.Second, 5*time.Second, 5),
+			"apbexternalroutenamespaces",
+		),
+		updatePolicyStatusFunc: updatePolicyStatusFunc,
 	}
 
 	return &m
+}
+
+func (m *externalPolicyManager) Run(wg *sync.WaitGroup, threadiness int) error {
+	defer utilruntime.HandleCrash()
+	klog.V(4).Info("Starting Admin Policy Based Route Controller")
+
+	_, err := m.namespaceInformer.AddEventHandler(
+		factory.WithUpdateHandlingForObjReplace(cache.ResourceEventHandlerFuncs{
+			AddFunc:    m.onNamespaceAdd,
+			UpdateFunc: m.onNamespaceUpdate,
+			DeleteFunc: m.onNamespaceDelete,
+		}))
+	if err != nil {
+		return err
+	}
+
+	_, err = m.podInformer.AddEventHandler(
+		factory.WithUpdateHandlingForObjReplace(cache.ResourceEventHandlerFuncs{
+			AddFunc:    m.onPodAdd,
+			UpdateFunc: m.onPodUpdate,
+			DeleteFunc: m.onPodDelete,
+		}))
+	if err != nil {
+		return err
+	}
+	_, err = m.routeInformer.AddEventHandler(
+		factory.WithUpdateHandlingForObjReplace(cache.ResourceEventHandlerFuncs{
+			AddFunc:    m.onPolicyAdd,
+			UpdateFunc: m.onPolicyUpdate,
+			DeleteFunc: m.onPolicyDelete,
+		}))
+	if err != nil {
+		return err
+	}
+
+	for i := 0; i < threadiness; i++ {
+		for _, workerFn := range []func(*sync.WaitGroup){
+			// processes route policies
+			m.runPolicyWorker,
+			// detects gateway pod changes and updates the pod's IP and MAC in the northbound DB
+			m.runPodWorker,
+			// detects namespace changes and applies polices that match the namespace selector in the `From` policy field
+			m.runNamespaceWorker,
+		} {
+			wg.Add(1)
+			go func(fn func(*sync.WaitGroup)) {
+				defer wg.Done()
+				wait.Until(func() {
+					fn(wg)
+				}, time.Second, m.stopCh)
+			}(workerFn)
+		}
+	}
+
+	// add shutdown goroutine waiting for m.stopCh
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+		// wait until we're told to stop
+		<-m.stopCh
+
+		m.podQueue.ShutDown()
+		m.routeQueue.ShutDown()
+		m.namespaceQueue.ShutDown()
+	}()
+
+	return nil
+}
+
+func (m *externalPolicyManager) runPolicyWorker(wg *sync.WaitGroup) {
+	for m.processNextPolicyWorkItem(wg) {
+	}
+}
+
+func (m *externalPolicyManager) processNextPolicyWorkItem(wg *sync.WaitGroup) bool {
+	wg.Add(1)
+	defer wg.Done()
+
+	key, shutdown := m.routeQueue.Get()
+
+	if shutdown {
+		return false
+	}
+
+	defer m.routeQueue.Done(key)
+
+	klog.V(4).Infof("Processing policy %s", key)
+	gwIPs, err := m.syncRoutePolicy(key.(string))
+	if err != nil {
+		klog.Errorf("Failed to sync APB policy %s: %v", key, err)
+	}
+
+	if m.updatePolicyStatusFunc != nil {
+		statusErr := m.updatePolicyStatusFunc(key.(string), gwIPs, err)
+		if statusErr != nil {
+			klog.Warningf("Failed to update AdminPolicyBasedExternalRoutes %s status: %v", key, statusErr)
+		}
+	}
+
+	if err != nil {
+		if m.routeQueue.NumRequeues(key) < maxRetries {
+			klog.V(4).Infof("Error found while processing policy %s: %v", key, err)
+			m.routeQueue.AddRateLimited(key)
+			return true
+		}
+		klog.Warningf("Dropping policy %q out of the queue: %v", key, err)
+		utilruntime.HandleError(err)
+	}
+	m.routeQueue.Forget(key)
+	return true
+}
+
+func (m *externalPolicyManager) onPolicyAdd(obj interface{}) {
+	_, ok := obj.(*adminpolicybasedrouteapi.AdminPolicyBasedExternalRoute)
+	if !ok {
+		utilruntime.HandleError(fmt.Errorf("expecting %T but received %T", &adminpolicybasedrouteapi.AdminPolicyBasedExternalRoute{}, obj))
+		return
+	}
+	key, err := cache.MetaNamespaceKeyFunc(obj)
+	if err != nil {
+		utilruntime.HandleError(fmt.Errorf("couldn't get key for object %+v: %v", obj, err))
+		return
+	}
+	klog.V(4).Infof("Adding policy %s", key)
+	m.routeQueue.Add(key)
+}
+
+func (m *externalPolicyManager) onPolicyUpdate(oldObj, newObj interface{}) {
+	oldRoutePolicy, ok := oldObj.(*adminpolicybasedrouteapi.AdminPolicyBasedExternalRoute)
+	if !ok {
+		utilruntime.HandleError(fmt.Errorf("expecting %T but received %T", &adminpolicybasedrouteapi.AdminPolicyBasedExternalRoute{}, oldObj))
+		return
+	}
+	newRoutePolicy, ok := newObj.(*adminpolicybasedrouteapi.AdminPolicyBasedExternalRoute)
+	if !ok {
+		utilruntime.HandleError(fmt.Errorf("expecting %T but received %T", &adminpolicybasedrouteapi.AdminPolicyBasedExternalRoute{}, newObj))
+		return
+	}
+
+	if oldRoutePolicy == nil || newRoutePolicy == nil {
+		utilruntime.HandleError(errors.New("invalid Admin Policy Based External Route provided to onPolicyUpdate()"))
+		return
+	}
+
+	if oldRoutePolicy.Generation == newRoutePolicy.Generation ||
+		!newRoutePolicy.GetDeletionTimestamp().IsZero() {
+		return
+	}
+
+	key, err := cache.MetaNamespaceKeyFunc(newObj)
+	if err != nil {
+		utilruntime.HandleError(fmt.Errorf("couldn't get key for object %+v: %v", newObj, err))
+	}
+	m.routeQueue.Add(key)
+}
+
+func (m *externalPolicyManager) onPolicyDelete(obj interface{}) {
+	_, ok := obj.(*adminpolicybasedrouteapi.AdminPolicyBasedExternalRoute)
+	if !ok {
+		tombstone, ok := obj.(cache.DeletedFinalStateUnknown)
+		if !ok {
+			utilruntime.HandleError(fmt.Errorf("couldn't get object from tomstone %#v", obj))
+			return
+		}
+		_, ok = tombstone.Obj.(*adminpolicybasedrouteapi.AdminPolicyBasedExternalRoute)
+		if !ok {
+			utilruntime.HandleError(fmt.Errorf("tombstone contained object that is not an Admin Policy Based External Route %#v", tombstone.Obj))
+			return
+		}
+	}
+	key, err := cache.MetaNamespaceKeyFunc(obj)
+	if err != nil {
+		utilruntime.HandleError(fmt.Errorf("couldn't get key for object %+v: %v", obj, err))
+		return
+	}
+	m.routeQueue.Add(key)
+}
+
+func (m *externalPolicyManager) onNamespaceAdd(obj interface{}) {
+	ns, ok := obj.(*v1.Namespace)
+	if !ok {
+		utilruntime.HandleError(fmt.Errorf("expecting %T but received %T", &v1.Namespace{}, obj))
+		return
+	}
+	if ns == nil {
+		utilruntime.HandleError(errors.New("invalid Namespace provided to onNamespaceAdd()"))
+		return
+	}
+	m.namespaceQueue.Add(ns)
+}
+
+func (m *externalPolicyManager) onNamespaceUpdate(oldObj, newObj interface{}) {
+	oldNamespace, ok := oldObj.(*v1.Namespace)
+	if !ok {
+		utilruntime.HandleError(fmt.Errorf("expecting %T but received %T", &v1.Namespace{}, oldObj))
+		return
+	}
+	newNamespace, ok := newObj.(*v1.Namespace)
+	if !ok {
+		utilruntime.HandleError(fmt.Errorf("expecting %T but received %T", &v1.Namespace{}, newObj))
+		return
+	}
+	if oldNamespace == nil || newNamespace == nil {
+		utilruntime.HandleError(errors.New("invalid Namespace provided to onNamespaceUpdate()"))
+		return
+	}
+	if oldNamespace.ResourceVersion == newNamespace.ResourceVersion || !newNamespace.GetDeletionTimestamp().IsZero() {
+		return
+	}
+	m.namespaceQueue.Add(newNamespace)
+}
+
+func (m *externalPolicyManager) onNamespaceDelete(obj interface{}) {
+	ns, ok := obj.(*v1.Namespace)
+	if !ok {
+		tombstone, ok := obj.(cache.DeletedFinalStateUnknown)
+		if !ok {
+			utilruntime.HandleError(fmt.Errorf("couldn't get object from tombstone %#v", obj))
+			return
+		}
+		ns, ok = tombstone.Obj.(*v1.Namespace)
+		if !ok {
+			utilruntime.HandleError(fmt.Errorf("tombstone contained object that is not a Namespace: %#v", tombstone.Obj))
+			return
+		}
+	}
+	if ns != nil {
+		m.namespaceQueue.Add(ns)
+	}
+}
+
+func (m *externalPolicyManager) runNamespaceWorker(wg *sync.WaitGroup) {
+	for m.processNextNamespaceWorkItem(wg) {
+
+	}
+}
+
+func (m *externalPolicyManager) processNextNamespaceWorkItem(wg *sync.WaitGroup) bool {
+	wg.Add(1)
+	defer wg.Done()
+
+	obj, shutdown := m.namespaceQueue.Get()
+
+	if shutdown {
+		return false
+	}
+
+	defer m.namespaceQueue.Done(obj)
+
+	err := m.syncNamespace(obj.(*v1.Namespace), m.routeQueue)
+	if err != nil {
+		if m.namespaceQueue.NumRequeues(obj) < maxRetries {
+			klog.V(4).Infof("Error found while processing namespace %s:%w", obj.(*v1.Namespace), err)
+			m.namespaceQueue.AddRateLimited(obj)
+			return true
+		}
+		klog.Warningf("Dropping namespace %q out of the queue: %v", obj.(*v1.Namespace).Name, err)
+		utilruntime.HandleError(err)
+	}
+	m.namespaceQueue.Forget(obj)
+	return true
+}
+
+func (m *externalPolicyManager) onPodAdd(obj interface{}) {
+	pod, ok := obj.(*v1.Pod)
+	if !ok {
+		utilruntime.HandleError(fmt.Errorf("expecting %T but received %T", &v1.Pod{}, obj))
+		return
+	}
+	if pod == nil {
+		utilruntime.HandleError(errors.New("invalid Pod provided to onPodAdd()"))
+		return
+	}
+	// if the pod does not have IPs AND there are no multus network status annotations found, skip it
+	if len(pod.Status.PodIPs) == 0 && len(pod.Annotations[nettypes.NetworkStatusAnnot]) == 0 {
+		return
+	}
+	m.podQueue.Add(pod)
+}
+
+func (m *externalPolicyManager) onPodUpdate(oldObj, newObj interface{}) {
+	o, ok := oldObj.(*v1.Pod)
+	if !ok {
+		utilruntime.HandleError(fmt.Errorf("expecting %T but received %T", &v1.Pod{}, o))
+		return
+	}
+	n, ok := newObj.(*v1.Pod)
+	if !ok {
+		utilruntime.HandleError(fmt.Errorf("expecting %T but received %T", &v1.Pod{}, n))
+		return
+	}
+	if o == nil || n == nil {
+		utilruntime.HandleError(errors.New("invalid Pod provided to onPodUpdate()"))
+		return
+	}
+	// if labels AND assigned Pod IPs AND the multus network status annotations are the same, skip processing changes to the pod.
+	if reflect.DeepEqual(o.Labels, n.Labels) &&
+		reflect.DeepEqual(o.Status.PodIPs, n.Status.PodIPs) &&
+		reflect.DeepEqual(o.Annotations[nettypes.NetworkStatusAnnot], n.Annotations[nettypes.NetworkStatusAnnot]) {
+		return
+	}
+	m.podQueue.Add(newObj)
+}
+
+func (m *externalPolicyManager) onPodDelete(obj interface{}) {
+	pod, ok := obj.(*v1.Pod)
+	if !ok {
+		tombstone, ok := obj.(cache.DeletedFinalStateUnknown)
+		if !ok {
+			utilruntime.HandleError(fmt.Errorf("couldn't get object from tombstone %#v", obj))
+			return
+		}
+		pod, ok = tombstone.Obj.(*v1.Pod)
+		if !ok {
+			utilruntime.HandleError(fmt.Errorf("tombstone contained object that is not a Pod: %#v", tombstone.Obj))
+			return
+		}
+	}
+	if pod != nil {
+		m.podQueue.Add(pod)
+	}
+}
+
+func (m *externalPolicyManager) runPodWorker(wg *sync.WaitGroup) {
+	for m.processNextPodWorkItem(wg) {
+	}
+}
+
+func (m *externalPolicyManager) processNextPodWorkItem(wg *sync.WaitGroup) bool {
+	wg.Add(1)
+	defer wg.Done()
+
+	obj, shutdown := m.podQueue.Get()
+
+	if shutdown {
+		return false
+	}
+
+	defer m.podQueue.Done(obj)
+
+	p := obj.(*v1.Pod)
+	err := m.syncPod(p, m.routeQueue)
+	if err != nil {
+		if m.podQueue.NumRequeues(obj) < maxRetries {
+			klog.V(4).Infof("Error found while processing pod %s/%s:%v", p.Namespace, p.Name, err)
+			m.podQueue.AddRateLimited(obj)
+			return true
+		}
+		klog.Warningf("Dropping pod %s/%s out of the queue: %s", p.Namespace, p.Name, err)
+		utilruntime.HandleError(err)
+	}
+
+	m.podQueue.Forget(obj)
+	return true
 }
 
 // getPoliciesForNamespaceChange returns a list of policies that should be reconciled because of a given namespace update.
